metodo che trova tutte le release di un repository in FilterRepository
